+++ doc/conf.py
@@ -2,8 +2,8 @@ import os
 
 # -- package specific configuration --
 project = "casm-bset"
-version = "2.0"  # The short X.Y version.
-release = "2.0.0"  # The full version, including alpha/beta/rc tags.
+version = "2.1"  # The short X.Y version.
+release = "2.1.0"  # The full version, including alpha/beta/rc tags.
 project_desc = "CASM cluster expansion basis set construction module"
 logo_text = "casm-bset"
 github_url = "https://github.com/prisms-center/CASMcode_bset/"